pyproject.toml
[tool.poetry]
name = "conference-crawler"
version = "0.1.0"
description = ""
authors = ["Johann Glock <johann.glocks@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
markdownify = "^0.11.2"
python-dateutil = "^2.8.2"
python-slugify = "^6.1.2"
Scrapy = "^2.6.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

